refactor: Home Time Schedule Card To Component
